Use scoped GIL unlock temporarily
