+++ cpp/arcticdb/pipeline/index_segment_reader.cpp
@@ -11,6 +11,7 @@
 #include <arcticdb/pipeline/index_segment_reader.hpp>
 #include <arcticdb/pipeline/slicing.hpp>
 #include <arcticdb/pipeline/index_fields.hpp>
+#include <arcticdb/pipeline/query.hpp>
 
 using namespace arcticdb::entity;
 using namespace arcticdb::stream;
@@ -99,6 +100,18 @@ bool IndexSegmentReader::is_pickled() const {
 
 bool IndexSegmentReader::has_timestamp_index() const {
     return tsd_.stream_descriptor().index().kind() == arcticdb::proto::descriptors::IndexDescriptor::Type::IndexDescriptor_Type_TIMESTAMP;
+}
+
+void check_column_and_date_range_filterable(const pipelines::index::IndexSegmentReader& index_segment_reader, const ReadQuery& read_query) {
+    util::check(!index_segment_reader.is_pickled()
+                    || (read_query.columns.empty() && std::holds_alternative<std::monostate>(read_query.row_filter)),
+                "The data for this symbol is pickled and does not support column stats, date_range, row_range, or column queries");
+    util::check(index_segment_reader.has_timestamp_index() || !std::holds_alternative<IndexRange>(read_query.row_filter),
+                "Cannot apply date range filter to symbol with non-timestamp index");
+    sorting::check<ErrorCode::E_UNSORTED_DATA>(index_segment_reader.tsd().stream_descriptor().sorted() == arcticdb::proto::descriptors::SortedValue::UNKNOWN ||
+                                                   index_segment_reader.tsd().stream_descriptor().sorted() == arcticdb::proto::descriptors::SortedValue::ASCENDING ||
+                                                   !std::holds_alternative<IndexRange>(read_query.row_filter),
+                                               "When filtering data using date_range, the symbol must be sorted in ascending order. ArcticDB believes it is not sorted in ascending order and cannot therefore filter the data using date_range.");
 }
 
 } // namespace  arcticdb::pipelines::index
